server/requirements.txt
flask
flask-cors
requests
ibm-watsonx-ai
python-dotenv